move functions from prof.go to mem.go. mem.go now prints the memory usage of individual API calls!
